Move fixtures
Following pytest's conventions, moved the fixtures to a seperate
conftest.py.
